pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "marketplace_search"
version = "0.1.0"
description = "Marketplace Search & Personalization Engine — Phase 0"
requires-python = ">=3.10"

dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "pyarrow>=12.0",        # Parquet I/O
    "pyyaml>=6.0",          # Config loading
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]
# Phase 1+ dependencies (not needed for Phase 0)
retrieval = [
    "scikit-learn>=1.3",
    "sentence-transformers>=2.2",
    "faiss-cpu>=1.7",
]
ranking = [
    "lightgbm>=4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"
